requirements.txt
flask==3.0.3
gunicorn==22.0.0
pandas==2.2.2
google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
